spec: Use array of string for custom default, plan modifier, and validator imports (#30)
Reference: https://github.com/hashicorp/terraform-plugin-codegen-spec/issues/28

When dealing with custom schema functionality, such as defaults, plan modifiers, and validators, the schema definition of these may require multiple code imports to successfully compile. This change adjusts the import specification of these from a single string to an array of strings. This also includes additional validation in the specification on those imports, leaving the `imports` property still optional, but if defined, it must contain at least one item and each item must not be an empty string.

As an example, the following specification definition wasn't possible before, but now is:

```json
          {
            "name": "list_attribute_default_custom",
            "list": {
              "computed_optional_required": "optional_computed",
              "default": {
                "custom": {
                  "imports": [
                    "github.com/hashicorp/terraform-plugin-framework/attr",
                    "github.com/hashicorp/terraform-plugin-framework/resource/schema/listdefault",
                    "github.com/hashicorp/terraform-plugin-framework/types"
                  ],
                  "schema_definition": "listdefault.StaticValue(types.ListValueMust(types.String, []attr.Value{types.StringValue(\"example\")}))"
                }
              },
              "element_type": {
                "string": {}
              }
            }
          },
```